# openapi-specifications/webhook/openapi-3.0.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "S2S api callback",
        "version": "1.0.0",
        "contact": {
            "name": "CpaaS Team"
        },
        "license": {
            "name": "Copy Right ALE"
        }
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Webhook server"
        }
    ],
    "paths": {
        "/connection": {
            "post": {
                "operationId": "connectionNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "ConnectionCtrl",
                "tags": [
                    "Connection"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbConnection"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/presence": {
            "post": {
                "operationId": "presenceNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "PresenceCtrl",
                "tags": [
                    "Presence"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbPresence"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/chat-state": {
            "post": {
                "operationId": "chatStateNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "ChatStateCtrl",
                "tags": [
                    "ChatState"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbChatState"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/receipt": {
            "post": {
                "operationId": "receiptNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "ReceiptCtrl",
                "tags": [
                    "Receipt"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbReceipt"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/all-receipt": {
            "post": {
                "operationId": "allReceiptNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "AllReceiptCtrl",
                "tags": [
                    "AllReceipt"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbAllReceipt"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/conversation": {
            "post": {
                "operationId": "conversationNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "ConversationCtrl",
                "tags": [
                    "Conversation"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbConversation"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/room-invite": {
            "post": {
                "operationId": "roomInviteNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "RoomInviteCtrl",
                "tags": [
                    "RoomInvite"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbRoomInvite"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/room-member": {
            "post": {
                "operationId": "roomMemberNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "RoomMemberCtrl",
                "tags": [
                    "RoomMember"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbRoomMember"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/room-state": {
            "post": {
                "operationId": "roomStateNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "RoomStateCtrl",
                "tags": [
                    "RoomState"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbRoomState"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/message": {
            "post": {
                "operationId": "messageNotificationReceived",
                "responses": {
                    "200": {
                        "description": "return this code if your server accepts the callback"
                    }
                },
                "x-swagger-router-controller": "MessageCtrl",
                "tags": [
                    "Message"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CbMessage"
                            }
                        }
                    },
                    "required": true
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CbConnection": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "connection": {
                        "type": "object",
                        "properties": {
                            "state": {
                                "type": "string",
                                "enum": [
                                    "ready",
                                    "terminated"
                                ]
                            }
                        },
                        "required": [
                            "state"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "connection"
                ],
                "title": "CbConnection"
            },
            "CbPresence": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "presence": {
                        "type": "object",
                        "properties": {
                            "from": {
                                "type": "string",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            },
                            "show": {
                                "type": "string",
                                "enum": [
                                    "away",
                                    "chat",
                                    "dnd",
                                    "xa"
                                ],
                                "example": "away",
                                "description": "presence 'show' attribute"
                            },
                            "status": {
                                "type": "string",
                                "example": "out for lunch",
                                "description": "presence 'status' attribute"
                            }
                        },
                        "required": [
                            "from"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "presence"
                ],
                "title": "CbPresence"
            },
            "CbChatState": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "chatstate": {
                        "type": "object",
                        "properties": {
                            "state": {
                                "type": "string",
                                "enum": [
                                    "composing",
                                    "paused",
                                    "inactive",
                                    "active"
                                ],
                                "description": "The chat state"
                            },
                            "conversation_id": {
                                "type": "string",
                                "example": "1555081424877876"
                            },
                            "peer": {
                                "type": "string",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            }
                        },
                        "required": [
                            "state",
                            "conversation_id",
                            "peer"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "chatstate"
                ],
                "title": "CbChatState"
            },
            "CbReceipt": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "receipt": {
                        "type": "object",
                        "properties": {
                            "msg_id": {
                                "type": "string",
                                "example": "cf05daec-c32e-11e8-affe-309c2364ed11"
                            },
                            "entity": {
                                "type": "string",
                                "enum": [
                                    "client",
                                    "server"
                                ],
                                "description": "The receipt emitter entity",
                                "example": "server"
                            },
                            "conversation_id": {
                                "type": "string",
                                "example": "1555081424877876"
                            },
                            "event": {
                                "type": "string",
                                "enum": [
                                    "received",
                                    "read"
                                ],
                                "example": "received"
                            }
                        },
                        "required": [
                            "msg_id",
                            "entity",
                            "event"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "receipt"
                ],
                "title": "CbReceipt"
            },
            "CbAllReceipt": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "all-receipt": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "enum": [
                                    "all-received",
                                    "all-sent"
                                ],
                                "description": "Receipt applies to 'sent' or 'received' messages",
                                "example": "all-received"
                            },
                            "conversation_id": {
                                "type": "string",
                                "example": "1555081424877876"
                            },
                            "with": {
                                "type": "string",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            }
                        },
                        "required": [
                            "id",
                            "conversation_id",
                            "with"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "all-receipt"
                ],
                "title": "CbAllReceipt"
            },
            "CbConversation": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "conversation": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "the conversation Id",
                                "example": "1554380274531685"
                            },
                            "action": {
                                "type": "string",
                                "enum": [
                                    "create",
                                    "update",
                                    "delete"
                                ],
                                "description": "the action type",
                                "example": "update"
                            },
                            "peer": {
                                "type": "string",
                                "description": "the peer Id",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "user",
                                    "bot",
                                    "room"
                                ],
                                "description": "the peer type",
                                "example": "user"
                            },
                            "mute": {
                                "type": "boolean",
                                "description": "the mute flag",
                                "example": "true"
                            }
                        },
                        "required": [
                            "id",
                            "action",
                            "peer",
                            "type"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "conversation"
                ],
                "title": "CbConversation"
            },
            "CbRoomInvite": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "room-invite": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "the room Id",
                                "example": "1554380274531685"
                            },
                            "by": {
                                "type": "string",
                                "description": "the inviting user Id",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            },
                            "name": {
                                "type": "string",
                                "description": "the room name",
                                "example": "Hot news"
                            }
                        },
                        "required": [
                            "id",
                            "by",
                            "name"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "room-invite"
                ],
                "title": "CbRoomInvite"
            },
            "CbRoomMember": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "room-member": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "the room Id",
                                "example": "1554380274531685"
                            },
                            "user": {
                                "type": "string",
                                "description": "the user Id",
                                "example": "59f1ecc4ea84f9b5c339958d"
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "accepted",
                                    "rejected",
                                    "unsubscribed"
                                ],
                                "description": "the member status",
                                "example": "accepted"
                            }
                        },
                        "required": [
                            "id",
                            "user",
                            "status"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "room-member"
                ],
                "title": "CbRoomMember"
            },
            "CbRoomState": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "the connection Id",
                        "example": "f8e81b4e-a7dc-11e8-a734-f894c2dba3d5"
                    },
                    "timestamp": {
                        "type": "string",
                        "description": "the message timestamp",
                        "example": "2018-08-23T08:52:04.001535Z"
                    },
                    "room-state": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "the room Id",
                                "example": "1554380274531685"
                            },
                            "event": {
                                "type": "string",
                                "enum": [
                                    "available",
                                    "unavailable",
                                    "deactivated",
                                    "resumed",
                                    "shutdown",
                                    "deleted"
                                ],
                                "description": "the room event",
                                "example": "deleted"
                            }
                        },
                        "required": [
                            "id",
                            "event"
                        ]
                    }
                },
                "required": [
                    "id",
                    "timestamp",
                    "room-state"
                ],
                "title": "CbRoomState"
            },
...
        }
    }
}